loads of changes
Restructured the code.

Added option to toggle allowing coalitions involving everyone else
Each coalition display now has a bar chart showing the makeup of the coalition, with a dashed line showing the target needed for a majority.
Coalitions are sorted by the size of the majority, then the leading party with the most seats, then by name.
Rewrote the algorithm to find coalitions.